+++ lib/src/event.dart
@@ -1,0 +1,2 @@
+// A marker interface for events.
+abstract class Event {}