+++ test/helpers/mock_event_callback.dart
@@ -1,0 +1,6 @@
+import 'package:mocktail/mocktail.dart';
+import 'package:skiff/src/event.dart';
+
+class MockEventCallback extends Mock {
+  dynamic call(Event e);
+}