+++ test/helpers/stub_event.dart
@@ -1,0 +1,3 @@
+import 'package:skiff/src/event.dart';
+
+class StubEvent extends Event {}